FEA Add hash encoding series method.
The method `hash_encode` has been added to the
`Series` class to allow the user to encode
the series' values into a bounded, non-negative
range using a hash function. Furthermore, the
user can also optionally combine the hashed
values with the hashed value of `Series.name` in order
to (most likely) obtain different encoded values
for the same original value in different columns.

A parameter had to be added to the `gdf_hash`
function to allow the caller to combine hashed
column values with a user specified ones.