FEA Add Series digitize method
Add `digitize` method to `Series` class that
mimics Numpy's `digitize` but does not
check for monotonicity of input bin values.

Expose `gdf_digitize` in `libcudf` to do the
work on the backend.

Transforming continuous valued features into
discrete ones is a common data processing step
in many machine learning models. The `digitize`
method allows the user to do just this if they
have, say, the histogram bin boundaries of a
continuous feature column.