Change null/valid counting logic in copy_if_else kernel to be more similar to the traditional method as in valid_if.
Flipped the logic in the kernel from counting nulls to counting valids, which makes the internal logic simpler and
more similar to other examples in cudf.

Switched to using an rmm::device_scalar for passing valid_count in/out of the kernel rather than hand allocating and
memcpying.

Added a number of new tests.