Rename some variables to not be quite as shorthand (w_ -> warp_) and eliminated some redundancy in cases where 'warp' is implied.
Fixed an issue where missing parentheses could result in us reading off the end of input data.

Added a special test case which sends in data that does not fit within the grid.  64 elements, grid size of 32.  This force
tests the loop indices warp_cur and index in the grid stride loop.