Try to create a `memoryview` of any data
Instead of merely check to see if a `memoryview` was passed to us, go
ahead and try to construct one from `data`. This should help coerce
anything that supports the Python Buffer Protocol (like `bytes` and
`bytearray`) into a `memoryview`, which we can easily construct a host
array from. Should be simple to create a device array from there.