Force to Python `int` first
Works around an issue where a CuPy scalar may be passed. Previously
`__array__` would be called by NumPy, but CuPy will throw there. So this
calls `int` on the object first. If it is a CuPy array, it will actually
coerce itself to a Python `int`. Then we can coerce it to a NumPy
`int32` value.