Use internal comparison for test
It appears this was using NumPy's `assert_allclose`. This has been here
since the very beginning of cuDF. While it may have made sense then, it
looks like the rest of the tests have moved on to this utility check. So
go ahead and do the same here to bring this test up-to-date.