Rewrite `StringColumn` to serialize internals
Instead of relying on the `nvstrings` API to build a `StringColumn` at
the other end, which may be expensive, go ahead and serialize the
internal columns correctly and rebuild them on the other side. This
avoids a bunch of unneeded copies when reconstructing the
`StringColumn`. Also it allows us to skip calling `nvstrings` until it
is actually needed.