Assert Dask CUDA serializers have `Buffer` frames
Previously this was being `assert`ed in the Dask serializers. However we
want to make sure we are passing `Buffer`s and not constructing other
objects (like Numba's `DeviceNDArrays`) in CUDA serializers as well. So
go ahead and move this to CUDA serializers where it still also affects
Dask serializers.