Check device/host objects are received as expected
When deserializing cuDF objects from Dask, make sure to check that
frames are on device (by checking for `__cuda_array_interface__`) when
`"cuda"` serialization is used and that frames are on host (by coercing
them to `memoryview`s) when `"dask"` serialization is used. If these
things don't work, make sure we raise an error. This would indicate the
data was not transmitted as expected, which would be a bug and would
indicate an unexpected performance degradation has occurred.