Add the serializer used to the `header`
Dask typically does this. So may overwrite the value we have here.
However just make sure we have set this value correctly. It makes it a
little easier to play with the serialization and deserialization
functions here. Plus if Dask doesn't set these for some reason, we have
handled it ourselves.