Pack/unpack cuDF frames during serialization
To cutdown on the number of frames that need to be transmitted over the
wire or transferred to/from host, pack all `Buffer`s into one
`DeviceBuffer` using CuPy to facilitate. In deserialization, use CuPy to
unpack them into a series of separate `DeviceBuffer` allocations.

Since cuDF already makes sure CuPy uses RMM for allocations, we need not
worry about this when creating `ndarray`s. However it is worth noting we
pay for a copy and a larger allocation both during serialization and
deserialization. It would be nice to avoid that, but it probably
requires some C++/CUDA code to do. So have held off on it for now.