Multiple binops fixes:
- Changed binops jit code to use std::common_type for casting operands to fix issues with
losing precision.

- Fixed CPU-side verification code (ASSERT_BINOP) to use epsilon checking for floating point
types instead of direct comparison.

- Fixed CPU-side binops operation code so that it takes TypeOut into account when using common_type. It
was only using the left and right operators so it missed the upcasting precision case.

- Changed the way BinaryOperationTest test fixture works for generating random numbers.  Previously
it was always using a generator of type uint8_t and then casting to output results.  This had the
quiet problem of always generating whole numbers, which was defeating a number of the floating point
tests.

- Added several new binops test to trigger precision-sensitive tests that were missing.

- Changed UniformRandomGenerator to use the 64 bit std mersenne twister random number generator.  Previously
it was using std::random_device, which was also a mersenne twister, however, it was using a random seed so
results between different runs of tests weren't repeatable.  Also changed it to use an incrementing seed
value for each new UniformRandomGenerator created - without this, every "randomly" generated column would
be the same values, which defeats most of the binops tests (for example, every mod test results in a value
of 0).