Rolling Window: Support DESC timestamps (c++ changes)
In grouped_timestamp_based_rolling_window(), it is possible that
the aggregated column is sorted in decreasing order of the accompanying
timestamp column. The current implementation assumes ascending order,
and returns the wrong results in this case.

This commit adds support to allow the caller to choose the sort-order
for the timestamp column.