Use RMM for Numba
When cuDF is imported, it configures Numba to make all allocations from
the RMM memory pool via RMMNumbaManager (similar to how the CuPy
allocator is set to be RMM).

References to RMM functions that re-implemented Numba functions are
removed and replaced with calls to the Numba version - since Numba is
now using RMM, the allocations will still come from the RMM pool.