Fix Numba deprecation warnings with Numba 0.49+
Many "internal" parts of Numba were moved with the release of 0.49.
The old module locations can still be imported with a shim for the 0.49
release, but its use results in the emission of NumbaDeprecationWarning
messages. The shims will be removed in Numba 0.50.

This PR avoids the deprecation warnings and enables import from the new
module paths where available, so that cuDF can be used with Numba 0.48,
0.49, or (a future release of) 0.50.

Once Numba 0.49 or above is a cuDF requirement, the `try .. except
ImportError` blocks can be replaced with imports exclusively from the
new locations.