Use `obj` underlying `memoryview` for pickling
Instead of creating a NumPy array to view the data in the `memoryview`,
just use the `obj` that already backs the `memoryview`. Currently this
is a NumPy array, so there is no semantic difference. Though this does
open the door to allowing other `obj`s that might back `memoryview`s to
be passed through in the future. For instance if/when RMM has a
`HostBuffer`, we could pass that through here as well. Meaning that
Python would then unpickle the result back into a `HostBuffer`. This
would be helpful when cutting down on device-to-host/host-to-device
transfer times.