Add nested type support to Java table serialization (#6705)
This adds support for LIST and STRUCT column types to JCudfSerialization.

In addition this contains fixes to problems in the Java nested column builder code and column equality testing code that were found with the jcudf nested type unit tests. A simpler syntax for adding test code to build tables with list and struct column elements was also added.