[REVIEW] Add dictionary support to cudf::minmax (#6764)
Reference #5963

Add support for dictionary column types in cudf::minmax. This PR also adds support for strings column types as well. The code was refactored to re-use the internal reduce_device function for all supported types. The strings specialization logic just creates a string_scalar result which requires a copy step to build a new string_scalar object from a string_view (output from the reduce_device function). The dictionary support was to just to substitute the keys() column in the detail function right before the type-dispatcher call.

Added a gtests for strings as well as dictionary with strings keys.