Verify that concatenating columns does not overflow size_type(#6809)
When concatenating large columns, it is possible to overflow the max size of a column (size_type).  We need to verify this won't happen ahead of time.   Previously, we were doing it for non-nested types, and only in the fused kernel concatenation case.   This addresses both issues.

Authors:
  - Dave Baranec <dbaranec@nvidia.com>

Approvers:
  - Paul Taylor
  - Ram (Ramakrishna Prabhu)
  - Christopher Harris

URL: https://github.com/rapidsai/cudf/pull/6809