Check output size overflow on strings gather(#6997)
Closes #6801 

This PR adds an extra reduce call in the libcudf gather specialization logic for strings column. This will check to make sure the output size of the gather does not exceed the size limit for the child characters column. The offsets column is first created with the individual output string sizes. Then the reduce call will add these sizes to check for overflow.

Also added a gtest to check for the overflow condition.

Authors:
  - davidwendt <dwendt@nvidia.com>

Approvers:
  - Devavret Makkar
  - Karthikeyan

URL: https://github.com/rapidsai/cudf/pull/6997