Handle nested string columns with no children in contiguous_split.(#6864)
Fixes a specific corner case:   String columns with no children (a special form of empty string column that can happen) that are nested inside a list (or struct) column.

This would be useful as a 0.17 PR but isn't strictly necessary, since it's pretty late.

Edit:  Updated the fix so that it always includes a record for src/dst buffers, even if they are of size 0 or have null data pointers.  The previous method that only checked the data pointer being null was unclean and didn't handle a particularly strange case that came up with the Spark plugin:  the plugin was reconstructing columns (on the receiver side of a shuffle) that had size 0 but a non-null data pointer.  This is technically legal but super weird.

Authors:
  - Dave Baranec <dbaranec@nvidia.com>
  - Karthikeyan <6488848+karthikeyann@users.noreply.github.com>

Approvers:
  - Karthikeyan (@karthikeyann)
  - Alfred Xu (@sperlingxx)
  - Karthikeyan (@karthikeyann)
  - Alfred Xu (@sperlingxx)
  - Karthikeyan (@karthikeyann)
  - Devavret Makkar (@devavret)

URL: https://github.com/rapidsai/cudf/pull/6864