Update doxyfile project number(#7161)
The Doxyfile project number is set to 0.16. I know I've seen it in the UI before but cannot find it now. I've updated the number just in case. And I've added a line to the update-version.sh (thanks @ajschmidt8 ) to automatically update the file when a new release is created.

Neither of these 2 files effect the CI/CD build.

Authors:
  - davidwendt <dwendt@nvidia.com>

Approvers:
  - Karthikeyan (@karthikeyann)
  - AJ Schmidt (@ajschmidt8)
  - Mark Harris (@harrism)

URL: https://github.com/rapidsai/cudf/pull/7161