Support contains() on lists of primitives (#7039)
Closes #6944.

This commit adds a method (`contains()`) to check whether each row of a `LIST` column contains the scalar value specified as an argument. The operation returns a `BOOL8` column (with as many rows as the input `LIST`), each row indicating `true` if the value is found, `false` if not.

Output `column[i]` is set to null if even one of the following holds true (in line with the semantics of `array_contains()` in SQL):
  1. The search key `skey` is null
  2. The list row `lists[i]` is null
  3. The list row `lists[i]` contains even *one* null, *and* `lists[i]` does not contain the search key.

This implementation currently supports the operation on lists of numerics or strings.

Authors:
  - MithunR (@mythrocks)

Approvers:
  - AJ Schmidt (@ajschmidt8)
  - Mark Harris (@harrism)
  - David (@davidwendt)
  - Karthikeyan (@karthikeyann)

URL: https://github.com/rapidsai/cudf/pull/7039