compile_udf: Cache PTX for similar functions
Compiling a UDF generated in a loop will result in a distinct
compilation for each loop iteration, because each new definition of the
UDF does not compare equal to any previous definition, and a new
compilation occurs. Furthermore, each new compilation returns PTX that
differs only in a trivial way (the generated code is the same but
function names are different), so JITify's cache also misses.

For example:

```python
for data_size in range(3):
    data = Series([3] * (2 ** data_size), dtype=np.float64)
    for i in range(3):
        data.applymap(lambda x: x + 1)
```

results in nine compilations when one would have sufficed.

This commit adds an additional cache to `compile_udf` keyed on the
signature, code, and closure variables of the function. This can hit for
distinct definitions of the same function. The existing `lru_cache`
wrapping `compile_udf` is left in place as it is expected to be able to
hash the function much more quickly, though I don't know if this has a
noticeable impact on performance - perhaps it would be worth removing it
for simplicity, so that there is only one level of caching.