Added in JNI support for out of core sort algorithm (#7381)
This makes changes to be able to support out of core sort on the Spark plugin. It adds sort order as an API so sorting that is not out of core does not need to gather columns that are only used for sort ordering.

It adds in versions of lower bound and upper bound that match more closely the APIs for sort so the sort order can be reused between them.

It updates the merge API to take an array of tables not just a list for simpler integration with Scala.

And makes a few minor changes to the sort order class so it can work better with Spark and debugging.

Authors:
  - Robert (Bobby) Evans (@revans2)

Approvers:
  - Jason Lowe (@jlowe)

URL: https://github.com/rapidsai/cudf/pull/7381