Fix java float/double parsing tests (#7473)
This is a part of #7410 and depends on it.  The java tests when they were first written, for what ever reason, did not use the correct string representation for several values including double max.  They ended up testing that the code was broken, which is not the right thing to do.  This updates those to have the correct values, but also has a comment about why the ideal value to test for overflow into `Inf` and `-Inf` is not ideal, but is a work around for small issues with the parsing code still.

Authors:
  - Robert (Bobby) Evans (@revans2)
  - Nghia Truong (@ttnghia)

Approvers:
  - Jason Lowe (@jlowe)

URL: https://github.com/rapidsai/cudf/pull/7473