Fix Java Parquet write after writer API changes (#7655)
This fixes the java build, but it required breaking changes to do it. I'll put up a corresponding change in the rapids plugin shortly.

https://github.com/rapidsai/cudf/issues/7654

was found as a part of this.

This is also not the final API that we will want. We need to redo how we configure the builders so that they can take advantage of the new APIs properly.

Authors:
  - Robert (Bobby) Evans (@revans2)

Approvers:
  - Jason Lowe (@jlowe)
  - Raza Jafri (@razajafri)

URL: https://github.com/rapidsai/cudf/pull/7655