Use dispatch_storage_type in get_data_ptr.
get_data_ptr was erroneously invoking
column_view::data<decimal32/64>().

As it happens, this ended up being okay
when the column did not have any offset as
column_view::data<int32>() and
column_view::data<decimal32>() will
return the same initial pointer when the
offset is zero. The problem occurs when you
attempt to advance this pointer, or if the
column had an offset (which internally will advance the pointer).