Struct hashing support for SerialMurmur3 and SparkMurmur3 (#7714)
Adding struct column support for serial Murmur3 and Spark-compatible Murmur3 hashing.  This explodes the struct column into the leaf columns before passing it to the existing hash support.  The validity of the parent struct columns can be ignored because hashing a null ends up as a no-op that returns the hash seed, so only the leaf columns within the struct column need to be considered for the hash computation.

Authors:
  - Jason Lowe (https://github.com/jlowe)

Approvers:
  - Robert (Bobby) Evans (https://github.com/revans2)
  - Conor Hoekstra (https://github.com/codereport)
  - Mark Harris (https://github.com/harrism)

URL: https://github.com/rapidsai/cudf/pull/7714