Don't install Thrust examples, tests, docs, and python files
Prune the amount of Thrust source tree that we install. On top
of reducing the size of cudf's install tree, this fixes an
issue where when using cudf via conda-build, conda presumes that
all python files inside cudf are valid and part of the package, and fails
to parse them.