Don't install Thrust examples, tests, docs, and python files (#7811)
Prune the amount of Thrust source tree that we install. On top of reducing the size of cudf's install tree, this fixes an
issue where when using cudf via conda-build, conda presumes that all python files inside cudf are valid and part of the package, and fails to parse them.

Authors:
  - Robert Maynard (https://github.com/robertmaynard)

Approvers:
  - Keith Kraus (https://github.com/kkraus14)

URL: https://github.com/rapidsai/cudf/pull/7811