Interval index and interval_range (#7182)
This PR is a follow up to PR #6984. It adds the IntervalIndex and interval_range classes to CuDF. IntervalIndex is needed for the `cut` and `histogram` methods and closes issue #5376.   

The main files to take note of in this PR are `index.py` and the two cudf `__init__.py` files. The rest of the file changes are related to PR6984 for creating the interval dtype.

Authors:
  - Marlene  (https://github.com/marlenezw)
  - Michael Wang (https://github.com/isVoid)

Approvers:
  - Keith Kraus (https://github.com/kkraus14)

URL: https://github.com/rapidsai/cudf/pull/7182