Java API change for supporting structs (#7730)
This is a very rough draft PR to tie down the interface change to support Structs for Parquet writer. Once we have the interface down, it's just a matter of coding in the rest of the pieces. 

Here is how I envision it to be used by the end-user. 
```
      ParquetWriterOptions options = ParquetWriterOptions.builder()
          .withColumnOptions(ParquetColumnWriterOptions.builder()
              .withColumnName("_c0")
              .withTimestampInt96(false)
              .withDecimalPrecision(4)
              .isNullable(true).build())
          .withStatisticsFrequency(ParquetWriterOptions.StatisticsFrequency.NONE)
          .withCompressionType(CompressionType.AUTO)
          .withMetadata("test", "test")
          .build();
```

I still don't know a good way to get rid of the old `withColumnNames` API on `ParquetWriterOptions`. We can't remove it as ORC is still using it. One option could be to just rip out the `WriterOptions` from the `ParquetWriterOptions` hierarchy.

Authors:
  - Raza Jafri (https://github.com/razajafri)

Approvers:
  - Mike Wilson (https://github.com/hyperbolic2346)
  - Robert (Bobby) Evans (https://github.com/revans2)

URL: https://github.com/rapidsai/cudf/pull/7730