Create a common code path for 1d Frames (#8115)
This PR unifies many code paths for 1-dimensional frame objects, namely Index and Series types. The bulk of the PR is moving code around, but there are a few renames (in particular `Index._values`->`Index._column`) that are necessary for unifying the APIs as well as some additional removals of unnecessary functions. The unification also fixes a few bugs (for instance, `bool(cudf.Index(...))` should raise an exception but wasn't) and adds a few missing APIs that were present in one class but not the other.

Authors:
  - Vyas Ramasubramani (https://github.com/vyasr)

Approvers:
  - https://github.com/brandon-b-miller
  - Michael Wang (https://github.com/isVoid)

URL: https://github.com/rapidsai/cudf/pull/8115