Fix bitmask_tests.cpp host accessing device memory (#8370)
This fixes an error in the `cleanEndWord` utility in the `bitmask_tests.cpp` source file. The code is trying to set the last remaining bits of a bitmask word by accessing the byte through a device pointer in host code. The end result was not actually being used so it appears the release compile optimized out the entire function. The expected segfault does occur in a debug build.
This PR corrects the logic to copy the last word, set the appropriate bits, and then copy the word back to the device buffer.

Authors:
  - David Wendt (https://github.com/davidwendt)

Approvers:
  - Mark Harris (https://github.com/harrism)
  - Nghia Truong (https://github.com/ttnghia)

URL: https://github.com/rapidsai/cudf/pull/8370