Remove special Index class from the general index class hierarchy (#8309)
Renames the old `Index` class to `BaseIndex` as the parent for all index types and creates a new `Index` class that subclasses `BaseIndex` to provide the expected API for `cudf.Index`. The critical change is that this class has a custom metaclass to make it look like `BaseIndex` from an inheritance perspective, and it takes on the `__new__` method that allows it to return different `Index` types depending on the input. This change will allow us to rewrite the hierarchy of different Index types in much simpler, more robust, and more performant ways. In particular, we will no longer have to override `__new__` in those classes, we will be able to make use of `__init__` rather than the custom `_initialize` method we are currently using, and we reduce ambiguities with respect to which and how many `__new__` and `__init__` methods are called.

Authors:
  - Vyas Ramasubramani (https://github.com/vyasr)

Approvers:
  - Ashwin Srinath (https://github.com/shwina)
  - GALI PREM SAGAR (https://github.com/galipremsagar)
  - Benjamin Zaitlen (https://github.com/quasiben)
  - Richard (Rick) Zamora (https://github.com/rjzamora)

URL: https://github.com/rapidsai/cudf/pull/8309