Add first-class dtype utilities (#8308)
This PR adds a new `cudf.api.types` module that aims to match `pandas.api.types` while providing the necessary compatibility layers for `cudf` objects that is missing from the corresponding `pandas` APIs. It also replaces most internal uses of `pandas.api.types` in an attempt to centralize all typing logic so that we have a single place in which to perform any special dtype handling as needed. This work is intended as a best-effort, first-pass attempt to isolate our dependence on pandas dtype APIs; while it resolves a number of incompatibilities with pandas and other unexpected behaviors, there are still a number of open questions that still need to be addressed to completely wrap this up. I've noted a number of TODOs in the code (for instance relating to our nested types or the different types of time types e.g. timedelta). Since getting all of these correct in a single PR will be almost impossible given the pervasive use of dtype utilities throughout our code, this PR is a good first step in that direction that I think we can merge and then work on incrementally fixing the outstanding issues rather than trying to perfect all our type handling here.

Authors:
  - Vyas Ramasubramani (https://github.com/vyasr)

Approvers:
  - Ashwin Srinath (https://github.com/shwina)

URL: https://github.com/rapidsai/cudf/pull/8308