Fix bug in replace_with_backrefs when group has greedy quantifier (#8575)
Closes #8569 

This essentially undoes the performance improvement made in #8460 since the logic mishandles a greedy quantifier pattern when it occurs inside an extract group. The internal regex logic is only able to track a single extract group when such a quantifier is specified. 

This PR does improve the interface for the internal extract call and adds some gtests for this issue.

Authors:
  - David Wendt (https://github.com/davidwendt)

Approvers:
  - Ram (Ramakrishna Prabhu) (https://github.com/rgsl888prabhu)
  - MithunR (https://github.com/mythrocks)

URL: https://github.com/rapidsai/cudf/pull/8575