Custom Sphinx Extension: `PandasCompat` (#8643)
This PR adds a custom sphinx extension `PandasCompat`. This extension adds two new rST directives to cudf docs code base.

```
.. pandas-compat::
.. pandas-compat-list::
```

The former creates a new section in cudf docstring specifically to demonstrate the deviation that cudf has to pandas. The second directive collects *all* these sections and compile as a separate list. For example, in `MultiIndex.get_loc`:

The new section is rendered as an `Admonition` tag.
![image](https://user-images.githubusercontent.com/13521008/124223660-f984ce00-dab8-11eb-957a-01bf652e7bad.png)

And there's a new separate page called `Pandas Compatibility Notes` that collects all notes in a single page.
![image](https://user-images.githubusercontent.com/13521008/124325434-7192d880-db39-11eb-9b53-c4abddc34026.png)


Note that we can introduce a link in this page to the original source.

Authors:
  - Michael Wang (https://github.com/isVoid)

Approvers:
  - GALI PREM SAGAR (https://github.com/galipremsagar)
  - Vyas Ramasubramani (https://github.com/vyasr)

URL: https://github.com/rapidsai/cudf/pull/8643