Fix `contains` check in string column (#8834)
Fixes: #8832 

This PR fixes `contains` check in the `StringColumn`.  We were using `f"^{item}$"` to generate a regex and do a `contains_re` to check for an exact match for `item` in the `StringColumn`, but this approach would break if `item` by itself has some regex special characters, so replaced these checks with `libcudf.search.contains` which does the exact check for `item` in the `StringColumn`.

Authors:
  - GALI PREM SAGAR (https://github.com/galipremsagar)

Approvers:
  - Ram (Ramakrishna Prabhu) (https://github.com/rgsl888prabhu)
  - Charles Blackmon-Luca (https://github.com/charlesbluca)

URL: https://github.com/rapidsai/cudf/pull/8834