Support bracket syntax for cudf::strings::replace_with_backrefs group index values (#8841)
Closes #8816 

The current `\d` syntax for the replacement template parameter will fail if a number immediately follows the index pattern as described in #8816. This PR adds support for the `${d}` pattern but only if the `\d` pattern is not found in the replacement string. This should minimize breaking any current templates already being used with this API.

Authors:
  - David Wendt (https://github.com/davidwendt)

Approvers:
  - Vukasin Milovanovic (https://github.com/vuule)
  - Christopher Harris (https://github.com/cwharris)

URL: https://github.com/rapidsai/cudf/pull/8841