Move compute_column API out of ast namespace (#8957)
The `compute_column` API is functionally analogous to `cudf::transform`, except that the former expresses the transformation in terms of a cudf expression while the latter accepts string UDFs to be JITted and evaluated. Moving `compute_column` into `transform.hpp` means that the only public APIs in the `ast` namespace are those required to _define_ an expression, so we clearly separate the logic for defining and parsing of expressions from APIs (like conditional joins and `compute_column`) that evaluate those expressions on device and then use the output in different ways.

Authors:
  - Vyas Ramasubramani (https://github.com/vyasr)

Approvers:
  - Jason Lowe (https://github.com/jlowe)
  - Vukasin Milovanovic (https://github.com/vuule)
  - Robert Maynard (https://github.com/robertmaynard)
  - Paul Taylor (https://github.com/trxcllnt)
  - Ray Douglass (https://github.com/raydouglass)

URL: https://github.com/rapidsai/cudf/pull/8957