Superimpose null masks for STRUCT columns.
Per #5700, when a STRUCT column is constructed, the null mask of the parent
column is bitwise-ANDed with that of all its children, such that a null row
in the parent column corresponds to nulls in all its children. This is done
recursively, allowing grand-child columns to also have nulls at the same
row positions.

`superimpose_parent_nulls()` makes this functionality available for columns
that might not have been constructed through `make_struct_column()`, e.g.
with columns received directly from Arrow. It does not require that the
`column_view` is modifiable. For a STRUCT `column_view` argument, a new
equivalent instance is created, with all its children's null masks modified
to account for the parent nulls.

`superimpose_parent_nulls()` can be used for all code that assumes that the
child null masks account for the nulls in the parents (and grandparents,
ad infinitum).