Support additional format specifiers in from_timestamps (#9047)
Reference #5991 

This PR adds support for the following format specifiers in `cudf::strings::from_timestamp`
```
%a and %A -- weekday names (passed into the API)
%b and %B -- month names (passed into the API)
%u - ISO weekday (1-7)
%w - weekday (0-6)
%U - week of the year (Sunday based)
%W - week of the year (Monday based)
%V - ISO week of the year
%G - Year based on ISO weeks
```

This adds a new parameter to the API for the caller to pass then string names for the weekdays and months. These are only required if the `%a, %b, %A, %B` specifiers are contained in the format string.

The change to `from_timestamps` is mainly a rewrite to include logic for these specifiers. Some common code required corresponding changes to `to_timestamps` and `is_timestamps` though these functions have not changed in this PR.

Authors:
  - David Wendt (https://github.com/davidwendt)

Approvers:
  - Devavret Makkar (https://github.com/devavret)
  - Conor Hoekstra (https://github.com/codereport)

URL: https://github.com/rapidsai/cudf/pull/9047