Remove unnecessary call to device_uvector::release() (#9237)
There was an unneeded call to `device_uvector::release()` that did not use the return value. rapidsai/rmm#857 marks that function `[[nodiscard]]`, so removing this allows libcudf to compile after the RMM PR is merged.

The `release()` is unnecessary because `prev_base_offsets` is passed by rvalue reference and therefore will be freed when the function exits (the uvector in the calling context it has been moved-from).

Authors:
  - Mark Harris (https://github.com/harrism)

Approvers:
  - Vyas Ramasubramani (https://github.com/vyasr)
  - Jake Hemstad (https://github.com/jrhemstad)

URL: https://github.com/rapidsai/cudf/pull/9237