Prevent Numba Performance Warnings in cuDF
Numba 0.54 introduces performance warnings whenever a kernel is launched
with low occupancy. Certain operations in cuDF can indirectly cause this
warning to be emitted, for example `DataFrame.iloc` (which one would not
expect to be particularly efficient anyway). This message could be
disconcerting for users and appear in a lot of workflows; therefore,
this commit configures Numba at cuDF import time to suppress the low
occupancy warning.

Fixes #9247.