Fix null count in statistics for parquet (#9303)
Fixes #9221 

The bug was in the method for calculating nulls `(num_rows - num_valid)` which needed `num_rows` to be summed in `typed_statistics_chunk` because it's not part of the untyped struct. This caused a problem because the serial `typed_statistics_chunk.reduce()` had the `num_rows` summation but the parallel `block_reduce(typed_statistics_chunk)` did not.

As part of the fix, `num_rows` is removed entirely and `null_count` is incremented in case of encountering nulls in the column.

Authors:
  - Devavret Makkar (https://github.com/devavret)

Approvers:
  - Vukasin Milovanovic (https://github.com/vuule)
  - Kumar Aatish (https://github.com/kaatish)
  - Ashwin Srinath (https://github.com/shwina)

URL: https://github.com/rapidsai/cudf/pull/9303