Push down parent nulls when flattening nested columns.
Fixes #9441.

`cudf::structs::detail::flatten_nested_columns()` is used to flatten `STRUCT`
columns, for use in operations where recursive descent is infeasible.
E.g. JOINing tables on a `STRUCT` join key.

If a `STRUCT` column's null-mask has not been superimposed over the member columns',
it is possible that a null in a `STRUCT` row is not reflected in the corresponding
member child/row. This causes the results of `flatten_nested_columns()` to cease
to be equivalent to the `STRUCT` column. The incorrect output proceeds to corrupt
the results of JOIN, as seen in #9441.

The erstwhile convention (i.e. to call `superimpose_parent_nulls()` before
`flatten_nested_columns()`) is easy to get wrong.

This commit ensures that `superimpose_parent_nulls()` is called in
`flatten_nested_columns()`. The now redundant calls (e.g. in `rank_scan`) have been
removed.