Fix regex handling of embedded null characters (#9470)
Closes #6196 

This fixes the regex pattern-matching logic to handle embedded null characters in a strings column like literals. The matching logic was previously identifying a 0-valued character (same as a null) as a flag to indicate the end of the string. A gtest is also included to test ignoring and matching on a null character.

Authors:
  - David Wendt (https://github.com/davidwendt)

Approvers:
  - Conor Hoekstra (https://github.com/codereport)
  - Bradley Dice (https://github.com/bdice)
  - Christopher Harris (https://github.com/cwharris)

URL: https://github.com/rapidsai/cudf/pull/9470