Use raw strings to avoid SyntaxErrors in parsed docstrings. (#9526)
I was looking at running cuDF Python tests with `-Werror` but ran into a few problems. One of those problems is that the docstrings in `string.py` use `\d` in code examples for regular expressions, which causes a SyntaxError if the docstring is parsed during test collection because `\d` is not a valid escape sequence in Python. Using a raw string in those methods' docstrings fixes the problem.

(Eventually I'd like to get our Python tests to fail if warnings are encountered -- with a list of excluded/filtered warnings that are outside our control.)

Authors:
  - Bradley Dice (https://github.com/bdice)

Approvers:
  - GALI PREM SAGAR (https://github.com/galipremsagar)

URL: https://github.com/rapidsai/cudf/pull/9526