Support re.Pattern object for pat arg in str.replace (#9573)
Closes #7904 

Depends on  #9439 

Adds support to accepting a pattern argument in cudf `str.replace` that is built using `re.compile`. The resulting `re.Pattern` has a member `pattern` that is the regex string that can be passed to libcudf. There is also `flags` member that can be used for libcudf APIs that accept a `regex_flags` parameter along with the `pattern` string.

Authors:
  - David Wendt (https://github.com/davidwendt)

Approvers:
  - https://github.com/brandon-b-miller
  - GALI PREM SAGAR (https://github.com/galipremsagar)

URL: https://github.com/rapidsai/cudf/pull/9573