Fail gracefully when compiling python UDFs that attempt to access columns with unsupported dtypes (#9359)
Depends on https://github.com/rapidsai/cudf/pull/9343

This PR updates the way that `apply` fails when the user tries to apply a function to the dataframe that uses a column with an unsupported dtype. As a side effect of moving towards a row like abstraction in user UDFs, it becomes tricky to determine if a field with an unsupported dtype is actually used. 

Consider the following example. This function should be allowed to execute, because even though the string field `b` is part of the "row", it is not used within the function.

```python
df = cudf.DataFrame({
    'a':[1,2,3],
    'b':['a','b','c'],
    'c':[7,8,9]
})

def f(row):
    return row['a'] + row['c']

res = df.apply(f, axis=1)
```

That said, the following should fail to be applied:

```python
def f(row):
    return row['a'] + row['b']

res = df.apply(f, axis=1)
```

In this case the information that we need to determine if we should throw can only be deduced from the actual function logic. Furthermore simplistic upfront analysis won't cut it in a lot of cases such as the following which should be allowed, since `x` is never used:

```python
def f(row):
    x = row['b']
    return row['a'] + row['c']
```

One can see that coupling this with variable aliases and the rest of python's language abstractions can quickly defeat simple algorithms for inspecting the function and determining if we should throw. The solution in this PR is to insert a special sentinel type in place of the type that would usually represent a supported dtype, and make it so that type deliberately can not be used for anything. That way during the function analysis that numba is already set up to do, it will fail and return a proper error to the user that explains what is happening if the field is used and if not, it will simply be compiled out of the final kernel.

Authors:
  - https://github.com/brandon-b-miller

Approvers:
  - Vyas Ramasubramani (https://github.com/vyasr)
  - Bradley Dice (https://github.com/bdice)
  - Graham Markall (https://github.com/gmarkall)

URL: https://github.com/rapidsai/cudf/pull/9359