Support `percent_rank()` aggregation
The `percent_rank()` aggregation is a ranking function, similar
to `rank()` and `dense_rank()`. It calculates the fractional/relative
rank of a row within a group of rows in a column, and returns a double
value in the range [0.0, 1.0] for each row in the group/column.

This commit includes the `libcudf` changes to support `percent_rank()`
aggregations, and the supporting JNI bindings.

Note that `percent_rank()` is typically used as a window aggregation. It
is implemented as a (grouped) scan aggregation, just as `rank()` and
`dense_rank()` because it operates across the whole group of rows. (i.e.
the window specification if fixed, spanning the entire group.)

References:
1. [SQL Server](https://docs.microsoft.com/en-us/sql/t-sql/functions/percent-rank-transact-sql)
2. [PostgreSQL](https://www.postgresql.org/docs/10/functions-window.html)
3. [SparkSQL](https://sparkbyexamples.com/spark/spark-sql-window-functions/#ranking-functions)