Add file size counter to cuIO benchmarks (#10154)
Most cuIO benchmarks used dataframes of fixed size as input. After writing to a file in the given format, its size can vary greatly depending on the encoding and compression.
This PR adds a counter to output the file size, as it can be often corelated with the performance of readers/writers.

Authors:
  - Vukasin Milovanovic (https://github.com/vuule)

Approvers:
  - Ram (Ramakrishna Prabhu) (https://github.com/rgsl888prabhu)
  - Mark Harris (https://github.com/harrism)

URL: https://github.com/rapidsai/cudf/pull/10154