Fix bitmask of the output for JNI of `lists::drop_list_duplicates` (#10210)
Previously, the Spark-rapids plugin only needs to call `lists::drop_list_duplicates` to create map which requires the input to be non-nullable. As such, the output of the JNI is just a lists column without bitmask. When operating on nullable input lists column, it produces incorrect results.

This PR fixes that.

Authors:
  - Nghia Truong (https://github.com/ttnghia)

Approvers:
  - Jason Lowe (https://github.com/jlowe)

URL: https://github.com/rapidsai/cudf/pull/10210