Java utilities to aid in accelerating aggregations on 128-bit types (#10201)
This adds a couple of custom kernels for Java to help accelerate sum aggregations on 128-bit types and check for overflows.  The first kernel extracts a 32-bit chunk from an 128-bit type which can be used to feed four 32-bit chunks into a sum aggregation.  The second kernel takes the resulting upscaled 64-bit integer results and reassembles the parts into a 128-bit type column along with a boolean column to indicate whether the value overflowed.

By splitting the 128-bit type into 32-bit chunks, a sum aggregation on DECIMAL128 which is a sort-based aggregation can be turned into a hash-based aggregation on 32-bit integer inputs for improved performance.  As a bonus, this approach can also check for overflow which is difficult to do when aggregating on DECIMAL128 sums directly.

Authors:
  - Jason Lowe (https://github.com/jlowe)

Approvers:
  - Jake Hemstad (https://github.com/jrhemstad)
  - Robert (Bobby) Evans (https://github.com/revans2)

URL: https://github.com/rapidsai/cudf/pull/10201