Allow Java bindings to use default decimal precisions when writing columns (#10276)
Closes #9851

This PR is to fix the bug raised in #9851 : Currently, Java bindings set decimal precision for each column when building OrcWriterOptions. It fills zero for non-decimal columns which do not carry precision information. In principle, we should only set precision for decimal columns.

It is not easy to write a test for this change, since we won't try to read non-decimal data as decimal type in both spark-rapids and cuDF Java.

Authors:
  - Alfred Xu (https://github.com/sperlingxx)

Approvers:
  - Jason Lowe (https://github.com/jlowe)
  - Nghia Truong (https://github.com/ttnghia)

URL: https://github.com/rapidsai/cudf/pull/10276