Add support for tdigest and merge_tdigest aggregations through cudf::reduce (#10433)
Previously, these aggregations only worked with groupby.  Now they can be invoked through `cudf::reduce`, producing scalar tdigest values (which under the hood are simply struct columns with 1 row).

The difference between the groupby and reduce versions is minimal.  They are both fundamentally `reduce_by_key` operations, where the keys represent the bucketing of input values to merged centroids.   In the case of `groupby`, the keys are further partitioned by the specific input group.   So the bulk of the changes are simply adding a few extra template parameters to various internal functions to allow the `reduce` path to behave as if it were just a constant group.

Similarly, many of the groupby tests which involved single groups have been refactored/repurposed for the reduce tests.

Most of the important changes are in `tdigest_aggregation.cu`

Authors:
  - https://github.com/nvdbaranec

Approvers:
  - Vyas Ramasubramani (https://github.com/vyasr)
  - Devavret Makkar (https://github.com/devavret)
  - Christopher Harris (https://github.com/cwharris)

URL: https://github.com/rapidsai/cudf/pull/10433