Fix temp data cleanup in `test_text.py` (#10524)
Noticed that when running pytests, cuDF was leaving a temporary file used for a specific test in the testing data folder without removing it. I think  our tests mostly write to `tmp` to avoid this issue.

In switching it to point to `tmp` I noticed the intended data is not being actually written to the file, instead it was writing the string `__repr__` of the generator. I also shrunk the data size by many orders of magnitude, because in the intended case where the data is meant to be truly generated, 30 mil ends up being a little cumbersome and slow.

Authors:
  - https://github.com/brandon-b-miller

Approvers:
  - GALI PREM SAGAR (https://github.com/galipremsagar)

URL: https://github.com/rapidsai/cudf/pull/10524