Reduce kernel calls to build strings findall results (#10559)
Changes implementation of `cudf::strings::findall` to build the intermediate results with a single kernel call instead of once per output column. The number of output columns is dependent on the string with the most matches. The result is a table of columns where each row is the matches for the corresponding input string. So a 2D vector is created to hold the results and padded with nulls as appropriate. This way, the regex engine is now only called once per match per string.

Each output column still requires a gather to build the output table. Although this requires more working memory to hold the intermediate results, the performance gain measured by the current benchmarks shows a 10-15% improvement.

The tests were not changed since no functional behavior was added or modified.

Authors:
  - David Wendt (https://github.com/davidwendt)

Approvers:
  - Mike Wilson (https://github.com/hyperbolic2346)
  - Nghia Truong (https://github.com/ttnghia)
  - Karthikeyan (https://github.com/karthikeyann)

URL: https://github.com/rapidsai/cudf/pull/10559