Refactor cudf::strings::count_re API to use count_matches utility (#10580)
Refactors the `cudf::strings::detail::count_re` function to reuse the `cudf::strings::detail::count_matches` utility created for findall, extractall, and split. The kernel code was identical with the only main difference the size of the output column. So the output size was added as a parameter to `count_matches` and the callers appropriately updated.

Authors:
  - David Wendt (https://github.com/davidwendt)

Approvers:
  - Bradley Dice (https://github.com/bdice)
  - Vyas Ramasubramani (https://github.com/vyasr)

URL: https://github.com/rapidsai/cudf/pull/10580