Persist string statistics data across multiple calls to orc chunked write (#10694)
This is the second half of the chunked orc write statistics work. This part enables persisting the string data between write calls, building the file-level statistics from the stripe data, and writing out the statistics in a chunked-write file. Care is made to ensure that everything is persisted by re-using the same variable in the added test for both writes to ensure nothing is missed. This was verified to invalidate the first table before the second call to write.

~This will clean up once 10567 goes in as this is branched off that work.~

depends on #10567 
closes #5826

Authors:
  - Mike Wilson (https://github.com/hyperbolic2346)

Approvers:
  - Vyas Ramasubramani (https://github.com/vyasr)
  - MithunR (https://github.com/mythrocks)
  - Vukasin Milovanovic (https://github.com/vuule)

URL: https://github.com/rapidsai/cudf/pull/10694