Add NumPy to intersphinx references. (#10809)
This PR adds `numpy` to the list of intersphinx references, which is needed to cross-reference numpy functions in cuDF's documentation. The Python intersphinx reference was updated to resolve a warning caused by a redirect. This also fixes a reference to `numpy.asarray`, which is a function (not a method).

Authors:
  - Bradley Dice (https://github.com/bdice)

Approvers:
  - GALI PREM SAGAR (https://github.com/galipremsagar)

URL: https://github.com/rapidsai/cudf/pull/10809