Remove deprecated Index methods from Frame (#11073)
This PR moves a number of methods from Frame to IndexedFrame that are no longer supported for Index objects. The one holdout is `Frame.replace`, which is also deprecated for Index but is used internally by Index methods in one place. In order to make this PR as simple to review as possible, I have omitted that function here. All changes in this PR are trivial moves of code aside from the deletions of the warnings.

Authors:
  - Vyas Ramasubramani (https://github.com/vyasr)

Approvers:
  - Bradley Dice (https://github.com/bdice)

URL: https://github.com/rapidsai/cudf/pull/11073