Add new `_from_data_like_self` factory (#11140)
This PR introduces a small new utility function `_from_data_like_self` to provide an additional hook for `Frame` and `IndexedFrame` to customize `_from_data`. Specifically in instances where the outputs need to share the index of the input for `IndexedFrame` methods, this hook provides the necessary entrypoint. This method is analogous to the already existing `_from_columns_like_self` method.

In the past I made efforts to consolidate our factory methods, and I'm not super happy with proliferating more of them again. However, I think in this case it is worth the tradeoff since this gets us very close to our goal of removing the `_index` attribute from the `Frame` class. Moreover, in the intermediate term I expect both this method and `_from_data` to be subsumed by `_from_columns` when we eventually transition all of our internals (especially the `ColumnAccessor`) to store a list rather than a dict of columns.

Authors:
  - Vyas Ramasubramani (https://github.com/vyasr)

Approvers:
  - Bradley Dice (https://github.com/bdice)

URL: https://github.com/rapidsai/cudf/pull/11140