Feature axes (#10979)
This PR closes #9592 by adding axes functionality to cudf. Return a list representing the axes of DataFrame and Series. For dataframes, it returns a list of two elements: element zero is the row index and element one is the columns. For series, it returns a list containing the row index.

Authors:
  - https://github.com/VamsiTallam95

Approvers:
  - https://github.com/brandon-b-miller
  - Vyas Ramasubramani (https://github.com/vyasr)

URL: https://github.com/rapidsai/cudf/pull/10979