Relax arrow pinning to just 8.x and remove cuda build dependency from cudf recipe (#11412)
Arrow 8.0.1 just got built and uploaded to conda-forge and there's an associated PR to update the pinning to 8.0.1. Currently the recipes pin to 8.0.0 explicitly which will cause solving issues. Instead of just changing to 8.0.1, this PR relaxes the pinning to allow minor and patch versions to get pulled as there hasn't been any nvcc compiler issues in a while and Arrow strives to not introduce API / ABI breakages in minor and patch versions. Additionally, the cudf recipe seems to be unnecessarily setting requiring a cuda build of pyarrow which should no longer be needed after #10995.

Would really like this in for 21.08 if possible, but happy to change to just pin 8.0.1 if preferred.

Authors:
  - Keith Kraus (https://github.com/kkraus14)

Approvers:
  - AJ Schmidt (https://github.com/ajschmidt8)

URL: https://github.com/rapidsai/cudf/pull/11412