Add `create_random_column` function to the data generator (#11490)
Adds an API to create a single random column, so users don't need to create a table even when a single column is required. The interface is the same as `create_random_table`, except that it only takes a single data type.

Authors:
  - Vukasin Milovanovic (https://github.com/vuule)

Approvers:
  - Karthikeyan (https://github.com/karthikeyann)
  - Nghia Truong (https://github.com/ttnghia)
  - Mike Wilson (https://github.com/hyperbolic2346)

URL: https://github.com/rapidsai/cudf/pull/11490