Support NTH_ELEMENT(0) and NTH_ELEMENT(-1) as hash aggregations
This change adds special cases for NTH_ELEMENT(0) (i.e. FIRST())
and NTH_ELEMENT(-1) (i.e. LAST()) groupby aggregations to be
computed as hash aggregations, instead of the regular sort-based
groupby.

This change should allow groupby aggregation requests
to avoid sorting the input projection on the groupby keys, if:
  1. All aggregations in the requests are among the hash-supported
     aggregations (E.g. MIN, MAX, SUM, COUNT, etc.).
  2. Any NTH_ELEMENT aggregations in the requests are specifically
     for NTH_ELEMENT(0) or NTH_ELEMENT(-1).

Avoiding sorting should speed up groupby queries that include
these NTH_ELEMENT special cases. This will be of specific relevance
for CUDF-dependent systems that have first class support for
`FIRST()` and `LAST()` aggregations.

This change does not address NTH_ELEMENT for (N != 0 && N != -1).
A solution for the generic case is likely more involved, and
left to a later date.

It appears that the extant tests in tests/groupby/nth_element_tests.cpp
already cover the special cases of NTH_ELEMENT(0)/NTH_ELEMENT(-1), with
and without null exclusion, and with nested types.

Signed-off-by: MithunR <mythrocks@gmail.com>