Update buffer interop tests for cupy 11
When performing equality tests with cupy objects, the implementation no
longer falls back to numpy.equal for objects which have type(foo).__eq__
is object.__eq__ and instead returns False (for __eq__) and True (for
__ne__). These tests were previously relying on this behaviour of
implicit coercion of Buffer objects to numpy arrays which is now gone.
So instead explicit round-trip into cupy arrays and check for equality
of the result.