Handle some zero-sized corner cases in dlpack interop (#11449)
Previously, we did not explicitly check for zero-sized inputs in to/from_dlpack.
cupy>=11 changes behaviour in how it provides information for zero-sized arrays
(strides of empty tensors were previously reported as one, but now are zero).
This broke some of the validity checking in `from_dlpack`. To handle such cases,
check the shape of the incoming tensor as well as strides.

Authors:
  - Lawrence Mitchell (https://github.com/wence-)
  - GALI PREM SAGAR (https://github.com/galipremsagar)
  - Benjamin Zaitlen (https://github.com/quasiben)

Approvers:
  - Ashwin Srinath (https://github.com/shwina)
  - Bradley Dice (https://github.com/bdice)
  - GALI PREM SAGAR (https://github.com/galipremsagar)
  - AJ Schmidt (https://github.com/ajschmidt8)

URL: https://github.com/rapidsai/cudf/pull/11449