Tell jitify_preprocess where to search for libnvrtc
On machines with multiple CUDA Toolkits installed it is possible
to have a mismatch between the version of `nvcc` used to compile
code and the version of `libnvrtc` used for the JIT code. This
generally occurs when `LD_LIBRARY_PATH` points to a different
version of the CUDA Toolkit.

We now explicitly specify what toolkit library directory to search
when JIT code during libcudf compilation.