part1: Simplify BaseIndex to an abstract class (#10389)
This PR is in response to @vyasr  comment, as partial fix for PR https://github.com/rapidsai/cudf/issues/9593 :
 
`BaseIndex `should be reduced as closely as possible to an abstract class. While there are a subset of APIs that truly make sense for all types of index objects, in almost all cases the optimal implementation for `RangeIndex `(and `MultiIndex`, for that matter) is very different from the implementation for `GenericIndex`. In addition, this change reduces cognitive load for developers by simplifying the inheritance hierarchy

Authors:
  - Sheilah Kirui (https://github.com/skirui-source)

Approvers:
  - Matthew Roeschke (https://github.com/mroeschke)
  - Vyas Ramasubramani (https://github.com/vyasr)

URL: https://github.com/rapidsai/cudf/pull/10389