Enable returning string data from UDFs used through `apply` (#11933)
This PR introduces the ability to return a string from a UDF used through `DataFrame.apply` or `Series.apply`. It provides all of the plumbing needed to run the function `lambda st: return st`, but does not provide any APIs that return strings such as `strip` or `upper` - these will be added in a series of followups. A cast from `string_view` to `udf_string` is provided that numba will call when attempting to return a `string_view` into a `udf_string` array.

Authors:
  - https://github.com/brandon-b-miller
  - David Wendt (https://github.com/davidwendt)

Approvers:
  - Vyas Ramasubramani (https://github.com/vyasr)
  - Bradley Dice (https://github.com/bdice)

URL: https://github.com/rapidsai/cudf/pull/11933