Floor division uses integer division for integral arguments
Previously (since c8b897ccd) floor division for all pairs of arguments
was implemented by promoting both arguments to double, dividing, and
then taking the floor.

This produces results when both operands are integral that sometimes
don't match the behaviour of normal floor division in any of the
target languages that cuDF supports. Specifically for int64 operands,
any time an operand is larger than 2^53, the conversion to double is
not injective and so information is lost.

To address this, implement SFINAE overloads for the FloorDiv device
functor that use standard C++ type promotion semantics for division
with mixed type operands. In the case of integral types, this is then
just standard division; if either type is floating, we perform
floating point division and then take the floor.

If the calling library's requirements for type promotion differ from
those of C++, it is the now the responsibility of the caller to
promote the operands to floor division before calling into libcudf.

Closes #12120.