Only exclude join keys that are indices from key columns
Previously, if any of the join keys were indices, we assumed that they
all were, and provided an empty set of key columns with matching names
in the left and right dataframe. This does the wrong thing for mixed
join keys (on a combination of index and normal columns), producing
more output columns than is correct. To avoid this, only skip matching
key names if they name indices.

Closes #11550.