Test that RAPIDS_NO_INITIALIZE means no cuInit
When RAPIDS_NO_INITIALIZE is set, importing cudf is not allowed to
create a CUDA context. This is quite delicate since calls arbitrarily
far down the import stack _might_ create one.

To spot such problems, build a small shared library that interposes
our own version of cuInit, and run a test importing cudf in a
subprocess with that library LD_PRELOADed. If everything is kosher, we
should not observe any calls to cuInit.

If one observes bad behaviour, the culprit can then be manually
tracked down in a debugger by breaking on our cuInit implementation.