Re-enable `Recently Updated` Check
This PR re-enables the `Recently Updated` check.

It was previously disabled because it was frequently getting stuck in the `Pending` state.

After investigating the logs, I determined that the job was getting killed prematurely due to an insufficient `timeout` value for the Lambda function that executes the code. Therefore, the Lambda function was being killed before the batch job was able to finish updating the statuses on all of the open PRs.

I increased the `timeout` value from 30 seconds to 15 minutes in the commit below which should resolve that problem.

- https://github.com/rapidsai/ops-bot/commit/4148bb03e47998d4c5b75cf0e35d676b9f7614b5

[skip ci]