Test that cuInit is not called when RAPIDS_NO_INITIALIZE is set
An alternate approach to that tried in #12361, here we just script GDB
and check if we hit a breakpoint in cuInit. When RAPIDS_NO_INITIALIZE
is set in the environment, merely importing cudf should not call into
the CUDA runtime/driver (i.e. no cuInit should be called). Conversely,
to check that we are scripting GDB properly, when we create a cudf
object, we definitely _should_ hit cuInit.