tests: Skip cuInit tests if cuda-gdb is not found or not working
Since #12545, we use cuda-gdb for scripting (since that is already
installed in the nvidia/cuda:-devel docker images) to check that
RAPIDS_NO_INITIALIZE ensured that cuInit is not called on import.

Unfortunately, the official cuda-gdb-11-2 package for Debian-based
systems does not correctly advertise all its dependencies (we need to
manually install libtinfo5 and libncursesw5). Consequently cuda-gdb
does not work if the base image rapids builds against is for CUDA
11.2.

To workaround this, build the cuda-gdb command as a fixture that is
appropriately marked in the cases where it is either not installed, or
else not working due to broken dependencies.