Move date to build string in `conda` recipe
Similarly to https://github.com/rapidsai/rmm/pull/1195, this PR moves the date from the version to the build string in the `cudf` `conda` recipes.

This prevents a bug from occurring where the Python builds fail because the date string it computes is different than the one computed by the C++ build, therefore causing the Python build to search for a C++ build artifact that doesn't exist.