Fix a leak in a test and clarify some test names (#12781)
Fix a small leak of host memory in a java unit test. Also updates some tests that verify that a double free is safe, but don't make it clear from the logs that the double free is expected to be there. This is so we don't have to spend too much time debugging if this double free is expected or not.

Authors:
  - Robert (Bobby) Evans (https://github.com/revans2)

Approvers:
  - Kuhu Shukla (https://github.com/kuhushukla)
  - Jim Brennan (https://github.com/jbrennan333)
  - Jason Lowe (https://github.com/jlowe)
  - Raza Jafri (https://github.com/razajafri)
  - Gera Shegalov (https://github.com/gerashegalov)

URL: https://github.com/rapidsai/cudf/pull/12781