Correctly handle scalar indices in Index.__getitem__
It is not sufficient to check for isinstance(i, int) since the index
may be a numpy type for which this check is False. Instead, invert the
condition and check if the return value from _get_elements_from_column
is a Column, in which case we should get an Index back.

Closes #12954.