Fix multi-block avro reader logic.
Per PR feedback and discussions, there was consensus that some tests
should be added that exercise the multi-block logic of our avro reader
implementation.

That resulted in the implementation of a new test that unified all
previous tests proposed in this PR: test_avro_reader_multiblock(),
which identified significant changes that needed to be made in order to
support a) multi-block avro files *at all*, and b) multi-block avro
files in conjunction with the new skip_rows and num_rows functionality.

This commit implements the required functionality to support multi-block
avro files *and* the new skip_rows and num_rows functionality.