Allow compilation with any GTest version 1.10+
GTest max support for `Types` was removed in 1.11, so we remove
the workarounds in cudf_gtest.

Since we need to support our custom `Types` and the GTest 1.11+
version rework the type_list_utilities to be generic and not
depend on specific traits.

Also corrected the `<<` overloads for GTest printing so that
they work with GTest 1.11.