JNI changes for range-extents in window functions.
This commit adds back-end JNI changes to support explicit range-extents for ranged window functions.
The change is analogous to the addition of `cudf::range_window_bounds::extent_type`, in `libcudf`.
This change will allow for `STRING` order-by columns for range window functions. It is required because
without it, it would be impossible to differentiate between `UNBOUNDED` and `CURRENT ROW` for a window
function over a `STRING` order-by column.