Fix integer overflow in partition scatter_map construction
Although the input partition map may be any valid integral type, the
intermediate scatter map should have the same type as a valid row
index (and the offsets histogram), namely `size_type`. Previously, the
scatter map was created with the same integer type as the partition
map, which can result in integer overflow, and incorrect results, when
the partition map is a narrow integral type and the input table has
more rows than the width of the type.

Closes #7513.