Optionally return compression statistics from ORC and Parquet writers (#13294)
Partially address #13252

Adds an option to ORC and Parquet writers to take a shared pointer to a `writer_compression_statistics` objects. This object is then updated with the compression statistics from the written file.
In the case of chunked writing, statistics object is updated after each file. If a writing operation fails, statistics are not modified.
One statistics object can be reused in multiple write calls, but it should not be used concurrently in multiple threads.

Authors:
  - Vukasin Milovanovic (https://github.com/vuule)

Approvers:
  - Robert Maynard (https://github.com/robertmaynard)
  - MithunR (https://github.com/mythrocks)
  - Nghia Truong (https://github.com/ttnghia)
  - Mike Wilson (https://github.com/hyperbolic2346)

URL: https://github.com/rapidsai/cudf/pull/13294