Correctly reorder and reindex scan groupbys with null keys (#13389)
Scan-based groupbys are massaged back into pandas (original dataframe)
order by a post-processing step. Previously, this did the wrong thing
if the grouping key contained null (or nan) keys. In this situation
dropna=True will cause libcudf to produce an output table that is
smaller than the input frame. To mimic pandas we need to expand this
output to the original frame size, inserting nulls in the missing rows
and reordering correctly.

Furthermore, the previous reordering code had an out-of-bounds memory
access when there were null keys, since we were asking to group a
column of the same length as the result, but the grouping object expects
columns of length of the original input (which is larger with
dropna=True and null keys).

To fix these issues, compute the reordering on a column of appropriate
size, and, if dropna is true and any of the key columns have nulls, go
down a more expensive reordering path that inserts nulls correctly by
reindexing the result.

- Closes #13349
- Closes #12055

Authors:
  - Lawrence Mitchell (https://github.com/wence-)

Approvers:
  - Ashwin Srinath (https://github.com/shwina)

URL: https://github.com/rapidsai/cudf/pull/13389