Ensure single_lane_block_sum_reduce is safe to call in a loop
Since shared memory declarations have a lifetime tied to the block
they are executed on (and not the function scope), we must sync the
block before returning the result to ensure that there are no
RAW/WAR-conflicts between successive invocations of
single_lane_block_sum_reduce within a kernel. Previously, it was
possible for the i+1th call to conflict with the ith call in a loop.

- Closes #13485.