Exposure Tracked Buffer (first step towards unifying copy-on-write and spilling) (#13307)
The first step towards unifying copy-on-write and spillable buffers. 

This PR re-implement copy-on-write by introducing a `ExposureTrackedBuffer` and `BufferSlice`. The idea is that when `copy-on-write` (and in a follow-up PR later, when `spill`) is enabled, we use `BufferSlice` throughout cudf. 
`BufferSlice` is a _view_ of a `ExposureTrackedBuffer` that implements copy-on-write semantics by tracking the number of `BufferSlice` that points to the same `ExposureTrackedBuffer`.

Authors:
  - Mads R. B. Kristensen (https://github.com/madsbk)

Approvers:
  - Vyas Ramasubramani (https://github.com/vyasr)
  - Lawrence Mitchell (https://github.com/wence-)

URL: https://github.com/rapidsai/cudf/pull/13307