Fix loc-getitem ordering when index contains duplicate labels
If the index has duplicate labels then it is not sufficient to sort
with the ordering induced by gathering the requested keys since the
duplicate keys arrive in non-deterministic order from the join in the
table that is being looked up from. To fix this, sort additionally by
a secondary key that is the ordering induced by gathering from the
donor table.

- Closes #13658