Fix combined filtering and column projection in `dask_cudf.read_parquet` (#13697)
This is the dask-cudf version of #13666, which fixes the case that the `filters` argument includes column names that are not included in the `columns` argument to `cudf.read_parquet`. It turns out that we need to add the exact same fix for the dask-specific `read_parquet` code path as well.  Note that it was just an oversight to leave this out of #13666 - This is currently a dask-expressions blocker.

Authors:
  - Richard (Rick) Zamora (https://github.com/rjzamora)

Approvers:
  - Lawrence Mitchell (https://github.com/wence-)

URL: https://github.com/rapidsai/cudf/pull/13697