Fix construction of DataFrames from dict when columns are provided
If a columns argument is provided to the dataframe constructor, this
should be used to select columns from the provided data dictionary.
The previous logic did do this correctly, but didn't preserve the
appropriate order of the resulting columns (which should come out in
the order that the column selection is in).

- Closes #13738