Workaround when Arrow has no S3 support.
On application shutdown libcudf needs to call `EnsureS3Finalized`
before exit. This is tricky since the arrow might not be compiled
with S3 support enabled.

So use dlopen to check for the symbol existence, and call when
it does exist. Only do this in the S3 tests for now as we figure
out a longer term solution.