Raise error when `astype(object)` is called in pandas compatibility mode (#13862)
This PR disables type-casting to `object` dtype for non-string columns. It's a no-op for string columns. We disable this only in pandas-compatibility mode since the usage of `np.dtype("O")` is wide-spread and not quite possible to remove it's support entirely at this point.

Authors:
  - GALI PREM SAGAR (https://github.com/galipremsagar)

Approvers:
  - Bradley Dice (https://github.com/bdice)

URL: https://github.com/rapidsai/cudf/pull/13862