Remove quadratic runtime due to accessing Frame._dtypes in loop
Frame._dtypes maps column names to dtypes, however, it is a property
that is computed on-demand. Consequently, a seemingly innocuous dict
lookup is actually O(N). When used in a loop over columns, this makes
an O(N) loop into an O(N^2) one.

This mostly bites on IO when reading data with many thousands of
columns. To fix this, manually move access of Frame._dtypes outside of
any loop over columns.

A more systematic way might be to make this a cached property, but the
cache invalidation is rather hard to reason about.

- Closes #14005