Reduce memory usage of as_categorical_column
The main culprit is in the way the codes returned from _label_encoding
were being ordered. We were generating an int64 column for the order,
gathering through the left gather map, and then argsorting, before
using that ordering as a gather map for the codes.

We note that gather(y, with=argsort(x)) is equivalent to
sort_by_key(y, with=x) so use that instead (avoiding an unnecessary
gather). Furthermore we also note that gather([0..n), with=x) is just
equivalent to x, so we can avoid a gather too.

This reduces the peak memory footprint of categorifying a random
column of 500_000_000 int32 values where there are 100 unique values
from 24.75 GiB to 11.67 GiB.