Fix assert failure for range window functions
Fixes #13855.

This commit fixes an erroneous assert check in range window functions,
where the wrong data-type is checked for timestamp order-by columns.

For timestamps, it's the duration::rep type that should be checked,
when accessing the order by column.

This fix allows ROLLING_TEST to complete correctly, in DEBUG mode.