Use call stack to record per-function call time
If a _fast_slow_function_call errors in the fast path, it may be that
the function name called in the "return" event doesn't match the one
recorded in the "call" event. This particularly occurs if the
user-facing call reaches some internal API in cudf that then produces
an error (for example inside setitem of some kind).

To avoid this, use a call stack to record the currently executing
fast-slow call and pop the stack in the "return" event.

- Closes rapidsai/xdf#369