Give final proxy types module of slow type
Dask and Ibis have a lazy dispatching mechanism that registers based
on type names (not types themselves). When a type dispatch occurs,
this lazy registration kicks in and checks to see if the type's module
name is one that it knows about. For this to work with our pandas
wrappers the final type wrappers (DataFrame, Series, ...) must
advertise that they come from a module that begins with "pandas."

To achieve this, rewrite the __module__ attribute so that this is
true. Since the __module__ should be the place that we can import the
type from, we use the module corresponding to the slow type (since
this really is where we can import the type from!).

- Closes rapidsai/xdf#370