"Intermediate" types from pandas namespace are proxies
Previously, although accessing pandas.DataFrame would return a proxied
"final" type, the same was not true of "intermediate" types (such as
DataFrameGroupBy). For example
pandas.core.groupby.generic.DataFrameGroupBy was not proxied. This
defeats type dispatching in libraries such as Ibis that generate an
execution plan and dispatching for all intermediate types in that
plan. Consequently, although we can correctly run df.groupby(...) on
accelerated mode, Ibis does not know what to do with the result (which
is not a pandas.core.groupby.generic.DataFrameGroupBy like it
expects).

To fix this, update the accelerated module wrapped attributes with the
known intermediate proxies as well as the final ones. While we're
here, also give them the appropriate module name from the slow type.

- Closes rapidsai/xdf#377