Match pandas ordering for joins in pandas-compat mode
If we pass sort=True to merges we are on the hook to sort the result
in order with respect to the key columns. If those key columns have
repeated values there is still some space for ambiguity. Currently we
get a result back whose order (for the repeated key values) is
determined by the gather map that libcudf returns for the join. This
does not come with any ordering guarantees.

When sort=False, pandas has join-type dependent ordering guarantees
which we also do not match. To fix this, in pandas-compatible mode
only, reorder the gather maps according to the order of the input
keys. When sort=False this means that our result matches pandas
ordering. When sort=True, it ensures that (if we use a stable sort)
the tie-break for equal sort keys is the input dataframe order.

While we're here, switch from argsort + gather to sort_by_key when
sorting results.

- Closes #14001