cudf jit parser now supports .pragma instructions with quotes
During testing of CUDA 12.3 it was found that python user defined
functions could generate .pragma instructions that have quotes,
so as "nounroll" and "used_bytes_mask N".

The jit parser now makes sure to escape quotes that are part
of a `.pragma` instruction so that it can be properly passed
to nvrtc/jitify as a CUDA header.