Expunge as_frame conversions in Column algorithms
Previously a number of algorithms on Columns first converted to a
single column frame and called a frame-based algorithm (which calls
directly into libcudf using the column we first thought of). This is
unnecessary since we already have the column to hand when calling the
same algorithm at the column level. Moreover, in many cases where the
algorithm is a user-facing API, the frame-based approach does more
work (for example conversions and dtype matching).

By removing this round trip we reduce some (unnecessary) overhead, and
also make the memory footprint and behaviour of column-based methods
more transparent.

- Closes #13565