Remove extra total chars size calculation from cudf::concatenate (#14540)
Changes bounds check logic in `cudf::concatenate` to remove the upfront chars bytes calculation on strings columns. 
The calculation was thrown away and redone later (and more efficiently) during the call to `cudf::strings::detail::concatenate`. There does not seem to be a need to check this twice. A gtest was added to make sure the appropriate exception is thrown as expected.
Reference #14202

Authors:
  - David Wendt (https://github.com/davidwendt)

Approvers:
  - Vukasin Milovanovic (https://github.com/vuule)
  - Karthikeyan (https://github.com/karthikeyann)
  - Nghia Truong (https://github.com/ttnghia)

URL: https://github.com/rapidsai/cudf/pull/14540