Introduce new type matching scheme for join keys
This is much tighter than the current scheme, but makes sure that we
never accidentally promote in situations where we might produce false
positives.

The assumption is that most joins occur on columns with matching
dtypes. Where it is definitely safe to promote numeric types, we do
so, but never between kinds. For almost all other column types we
require exact dtype matching and now raise. The only exception is if
exactly one of the two columns is categorical in which case we
decategorize and go round again.