Add (implicit) handling for torch tensors in is_scalar
PyTorch tensors advertise that they support the number API, and hence
answer "True" to the question pd.api.types.is_scalar(torch_tensor).
This trips up some of our data ingest, since in as_index we check if
the input is a scalar (and raise) before handing off to as_column. To
handle this, if we get True back from pandas' is_scalar call,
additionally check that the object has an empty shape attribute (if it
exists).

See also:

- https://github.com/pytorch/pytorch/issues/99646
- https://github.com/pandas-dev/pandas/issues/52701