Drop python-snappy from dependencies. (#15161)
Previously `python-snappy` was a test dependency. It does not appear that we rely on this directly, as there are no instances of `import snappy`. Recently, pandas also dropped this dependency: https://github.com/pandas-dev/pandas/pull/54633

More generally, we can refactor the dependency list to use `pandas[all]` now that we require pandas 2.

Authors:
   - Bradley Dice (https://github.com/bdice)

Approvers:
   - Kyle Edwards (https://github.com/KyleFromNVIDIA)
   - Vyas Ramasubramani (https://github.com/vyasr)
   - GALI PREM SAGAR (https://github.com/galipremsagar)
   - Ray Douglass (https://github.com/raydouglass)