Performance optimizations for parquet sub-rowgroup reader. (#15020)
This PR implements a basket of optimizations for the parquet reader to bring non-chunked reads close  to par following the merge of the sub-rowgroup reader.  

The primary culprit for the performance hit was that in the case where we perform no splits, we were making a full copy of all of the pages into the subpass struct (including a pinned memory allocation).  This is unnecessary because we can just represent the pages in the subpass as a span that wraps the existing pages in the pass. 

In addition, several `hostdevice_vector`s used for work that could be done entirely device-side were converted to `rmm::device_uvector`.

Finally, I converted a number of functions that were taking hostdevice_vectors to use spans instead and added some missing operators to the `hostdevice_vector` class itself.

This PR doesn't recover all the time (there is some new work that we have to do in all cases) but it takes out most of the sting. A sample of some of the benchmarks that were most notably affected:

```
                       Original Time      Sub-rowgroup-implementation       This PR
parquet_read_decode
Int, device buffer 0   29260860778        26373181343                       28121328587
Int, device buffer 1   30692134492        27474241282                       29495189226

parquet_read_chunks
Int, device buffer     33895028252        29986276949                       32293548191
Float, device buffer   57055985251        49640274260                       55795392897
```

Authors:
  - https://github.com/nvdbaranec

Approvers:
  - Mike Wilson (https://github.com/hyperbolic2346)
  - Nghia Truong (https://github.com/ttnghia)

URL: https://github.com/rapidsai/cudf/pull/15020