[JNI] rmm based pinned pool (#15219)
Part of https://github.com/rapidsai/cudf/issues/14782.

This PR removes our old implementation of the java based pinned memory pool and replaces it with a jni layer on top of `rmm::pool_memory_resource<rmm::pinned_host_memory_resource>`

This PR does NOT set the default cuIO pinned host resource. That is happening after this PR goes in https://github.com/rapidsai/cudf/pull/15079. We'll need a follow on PR to change `PinnedMemoryPool.initialize` method to add an argument to set the cuIO pinned host resource.

I have run with this and version of it that are shared with cuIO and I can't find regressions in NDS at SF3K.

Note that we don't align anymore on our side. RMM is doing the same alignment we were doing before, using `std::max_align_t`.

Note also that the rmm pool doesn't have a quick way to find out what the current size is. So we had some tests that were asserting for this, and I have removed the asserts. If we would like to get that back I am happy to work with RMM to figure out how to do that.

Authors:
  - Alessandro Bellina (https://github.com/abellina)

Approvers:
  - Jim Brennan (https://github.com/jbrennan333)

URL: https://github.com/rapidsai/cudf/pull/15219