Roll back ipow changes due to register pressure. (#15242)
The addition of an array of integers in this function placed too much register pressure on our code base. This function is used by the fixed_point constructor and cast operators, so it potentially affects every kernel.  Too many unrelated kernels were impacted and suffered performance degradations to justify this change.  This reverts the algorithm introduced in #15110 to what it was previously, with some very minor tweaks.

Authors:
  - Paul Mattione (https://github.com/pmattione-nvidia)

Approvers:
  - Yunsong Wang (https://github.com/PointKernel)
  - Mike Wilson (https://github.com/hyperbolic2346)
  - Shruti Shivakumar (https://github.com/shrshi)
  - MithunR (https://github.com/mythrocks)

URL: https://github.com/rapidsai/cudf/pull/15242