Implement grouped product scan (#15254)
Although cumulative products are implemented for whole-frame scans, they were not for grouped aggregations.
Plumb through the necessary machinery to enable this. Only enabled for floating and integral types: the units make no sense for durations.

As for the whole-frame product aggregation, it is very easy to overflow the output type. For floating types this will result in `+/- inf` as the result. For signed integral types, behaviour is undefined on overflow.

- Closes #15253

Authors:
  - Lawrence Mitchell (https://github.com/wence-)
  - Bradley Dice (https://github.com/bdice)

Approvers:
  - David Wendt (https://github.com/davidwendt)
  - Bradley Dice (https://github.com/bdice)

URL: https://github.com/rapidsai/cudf/pull/15254