Handle case of scan aggregation in groupby-transform
When performing a groupby-transform with a scan aggregation, the
intermediate result obtained from calling groupby-agg is already the
correct shape and does not need to be broadcast to align with the
grouping keys.

To handle this, make sure that if the requested transform is a scan
that we don't try and broadcast.

While here, tighten up the input checking: transform only applies to a
single aggregation, rather than the more general interface offered by
agg.

- Closes #12621
- Closes #15448