cudf.pandas: Series dt accessor is CombinedDatetimelikeProperties
On the pandas Series type (not an instance) the dt attribute returns a
CombinedDatetimelikeProperties object, which advertises the attributes
of all possible datetime like dtypes. Previously we were proxying this
with a DatatimeProperties object, which doesn't advertise as many
properties. To allow wrapping libraries like dask that introspect the
object to work correctly, advertise like pandas on the type. The
instance still produces an object of the correct type due to dynamic
lookup and/or metaclass magic in cudf.pandas and pandas respectively.

- Closes #15522