strings::contains() for multiple search targets
This commit adds a new `strings::contains()` overload that allows
for the search of multiple scalar search targets in the same call.

The trick here is that a new kernel has been introduced, to extend
the "string-per-warp" approach to search for multiple search keys
in the same kernel.

This approach allows CUDF to potentially reduce the number of kernels
launched for `string::contains()` by a factor of `N`, if all the
search keys can be specified in the same call.  This helps reduce
the kernel-launch overheads for processes that do large numbers
of calls to `string::contains()`.

Signed-off-by: MithunR <mythrocks@gmail.com>