strings::contains() for multiple search targets
This commit adds a new `strings::contains()` overload that allows
for the search of multiple scalar search targets in the same call.

The trick here is that a new kernel has been introduced, to extend
the "string-per-warp" approach to search for multiple search keys
in the same kernel.

This approach allows CUDF to potentially reduce the number of kernels
launched for `string::contains()` by a factor of `N`, if all the
search keys can be specified in the same call.  This helps reduce
the kernel-launch overheads for processes that do large numbers
of calls to `string::contains()`.

Signed-off-by: MithunR <mythrocks@gmail.com>

Changed iteration order, for better cache performance.

More optimizations:

1. Removed calls to `thrust::fill()`. The bool values are now explicitly written in the kernel.
2. Switched host-device copy to use async.

Revert "More optimizations:"

This reverts commit c0e355c7be116ebd79bcdc6aa77c36af182e5ec6.

This commit was wrong: The thrust::fill() checks for empty target strings.
If removed, we'll need to check for empty target strings for every input string
row.
This was better done the old way.

More improvements:

1. Removed thrust::fill call. Setting values explicitly in the kernel.
2. Switched from using io::hostdevice_vector to rmm::device_uvector. The string_view allocation is tiny.

This has helped reduce the time spent in strings::contains().

For small strings, delegate to thread-per-string algo.