Access `self.index` instead of `self._index` where possible (#15781)
Since `index` is defined as 

```python
@property
def index(self):
    return self._index
```

Get and set to `self.index` when possible. Setting to `self.index` ensures that we may not be creating an invalid `IndexedFrame` with a `len(index) != len(columns)`.

There are times when still setting `self._index` was necessary because some data was being swapped "inplace" and validation needed to be avoided. (Hoping to avoid this pattern in the future)

Authors:
  - Matthew Roeschke (https://github.com/mroeschke)

Approvers:
  - Vyas Ramasubramani (https://github.com/vyasr)

URL: https://github.com/rapidsai/cudf/pull/15781