Revert "Fix docs for IO readers and strings_convert" (#15872)
Reverts rapidsai/cudf#15842

The files the original PR added documentation for appear to contain some
text that is problematic for the Sphinx parser to extract from doxygen.
My best guess is that it's something in a table, since parsing doxygen
tables via Breathe is something I know can be tricky. We didn't catch
this issue because [we currently only build the text docs in nightly
builds, not
PRs](https://github.com/rapidsai/cudf/blob/branch-24.08/ci/build_docs.sh#L49),
and this issue only arises in those text builds. We can revisit adding
these docs in 24.08. For the sake of correctness, I have added back
building text docs in PRs in this PR (see #14856 for context on the
removal).