Utilities for decimal <--> floating conversion (#15359)
These are some utilities used by the upcoming decimal <--> floating conversion PR.  This has been submitted separately from that PR in order to spread out the complexity for review.  These functions are not called by any code in this PR.  

One function is used to extract the components of the floating point number.  Another function is used to set a floating point's sign bit and add some additional powers of two.  These are done using integer and bit operations, which is much faster than using the built-in functions and bottle-necking on the FP64 pipeline.  The final function is used to count the # of significant bits in a number.

Authors:
  - Paul Mattione (https://github.com/pmattione-nvidia)
  - Lawrence Mitchell (https://github.com/wence-)

Approvers:
  - Mark Harris (https://github.com/harrism)
  - Bradley Dice (https://github.com/bdice)
  - Mike Wilson (https://github.com/hyperbolic2346)

URL: https://github.com/rapidsai/cudf/pull/15359