Document how to use cudf.pandas in tandem with multiprocessing
We need to arrange that cudf.pandas.install() is run on the workers,
this requires that we programmatically install the metapath loader in
our script. Unfortunately, passing an initializer function to the pool
startup is not sufficient if any part of the script transitively loads
pandas at the top level.

- Closes #15246