Fix exclude regex in pre-commit clang-format hook
The clang-tidy changes in #15894 introduce a new exclude regex list to
the pre-commit clang-format hook. However, it was a single character
too long, ending with a |. Consequently, the exclude regex matched the
empty string, and hence excluded every C++ file.

Fix this, and apply formatting changes to the files that were modified
in the interim and were not clang-format compatible.