Fix initialization error in to_arrow for empty string views
When converting an empty string view to arrow, we don't bother with
copies from device, but rather create the arrow arrays directly. The
offset buffer is therefore a singleton int32 array with zero in it.

Previously, the initialization of this array was incorrect, since
mutable_data() returns a uint8_t pointer, and so setting the single
element could leave 24 or the 32 bits uninitialized.

Fix this by using memset instead to zero out the full buffer.