Remove decimal/floating 64/128bit switches due to register pressure (#16287)
The decimal <--> floating conversion PR reduced the performance of some of the AST and BINARYOP kernels due to register pressure.  This removes the switches that are the primary source of the register pressure, falling back to the old ipow() method for 64bit and 128bit integers.

Authors:
  - Paul Mattione (https://github.com/pmattione-nvidia)

Approvers:
  - Jayjeet Chakraborty (https://github.com/JayjeetAtGithub)
  - Nghia Truong (https://github.com/ttnghia)

URL: https://github.com/rapidsai/cudf/pull/16287