Avoid deprecation warning in internal to_arrow code
In #16297, we deprecated the use of `to_arrow` in favour of
`to_arrow_host` and `to_arrow_device`. However, the scalar detail
overload of `to_arrow` used the public table overload. So we get a
warning when compiling internal libcudf code. Fix this by using the
detail API, and fix a bug along the way where we were not passing
through the arrow memory resource.