Enable prefetching in cudf.pandas.install() (#16439)
This PR enables `cudf.pandas` managed memory prefetching in
`cudf.pandas.install()`, to ensure that prefetching is enabled for all
methods of enabling `cudf.pandas`.

I also fixed a bug in libcudf's prefetching logic, where it tried to
compute the number of characters in a strings column view even if the
string column view's data is `nullptr`. This errors, so we must avoid
the `chars_size()` call and stop the prefetch attempt early.