Ensure comparisons with pyints and integer series always succeed
When Python integers are compared to a series of integers, the result
can always be correctly defined no matter the values of the Python
integer.

This was always a very mild issue.  But with NumPy 2 behavior not
upcasting the computation result type based on the value anymore,
even things like:
```
cudf.Series([1, 2, 3], dtype="int8") < 1000
```
would fail.

N.B. NumPy/pandas also support exact comparisons when mixing e.g.
uint64 and int64.  This is another rare exception that cudf currently
does not support.